document-search/pyproject.toml
[tool.poetry]
name = "document-search"
version = "0.1.0"
description = ""
authors = ["a1omax <maksimevstratev@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"

aiohappyeyeballs = "2.4.6"
aiohttp = "3.11.12"
aiosignal = "1.3.2"
annotated-types = "0.7.0"
anyio = "4.8.0"
asgiref = "3.8.1"
attrs = "25.1.0"
backoff = "2.2.1"
bcrypt = "4.2.1"
beautifulsoup4 = "4.13.3"
build = "1.2.2.post1"
cachetools = "5.5.1"
certifi = "2025.1.31"
charset-normalizer = "3.4.1"
chroma-hnswlib = "0.7.6"
chromadb = "0.6.3"
click = "8.1.8"
colorama = "0.4.6"
coloredlogs = "15.0.1"
dataclasses-json = "0.6.7"
datasets = "3.3.0"
deepeval = "2.3.7"
Deprecated = "1.2.18"
dill = "0.3.8"
dirtyjson = "1.0.8"
distro = "1.9.0"
docx2txt = "0.8"
durationpy = "0.9"
execnet = "2.1.1"
faiss-cpu = "1.10.0"
fastapi = {extras = ["standard"], version = "^0.115.12"}
filelock = "3.17.0"
filetype = "1.2.0"
flatbuffers = "25.2.10"
frozenlist = "1.5.0"
fsspec = "2024.12.0"
google-auth = "2.38.0"
googleapis-common-protos = "1.67.0"
greenlet = "3.1.1"
groq = "0.18.0"
grpcio = "1.67.1"
h11 = "0.14.0"
httpcore = "1.0.7"
httptools = "0.6.4"
httpx = "0.28.1"
httpx-sse = "0.4.0"
huggingface-hub = "0.28.1"
humanfriendly = "10.0"
idna = "3.10"
importlib_metadata = "8.5.0"
importlib_resources = "6.5.2"
iniconfig = "2.0.0"
Jinja2 = "3.1.5"
jiter = "0.8.2"
joblib = "1.4.2"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
kubernetes = "32.0.0"
langchain = "0.3.18"
langchain-chroma = "0.2.2"
langchain-community = "0.3.17"
langchain-core = "0.3.35"
langchain-groq = "0.2.4"
langchain-openai = "0.3.5"
langchain-text-splitters = "0.3.6"
langsmith = "0.3.8"
llama-cloud = "0.1.12"
llama-cloud-services = "0.6.1"
llama-index = "0.12.17"
llama-index-agent-openai = "0.4.5"
llama-index-cli = "0.4.0"
llama-index-core = "0.12.17"
llama-index-embeddings-openai = "0.3.1"
llama-index-indices-managed-llama-cloud = "0.6.4"
llama-index-llms-openai = "0.3.19"
llama-index-multi-modal-llms-openai = "0.4.3"
llama-index-program-openai = "0.3.1"
llama-index-question-gen-openai = "0.3.0"
llama-index-readers-file = "0.4.5"
llama-index-readers-llama-parse = "0.4.0"
llama-parse = "0.6.1"
markdown-it-py = "3.0.0"
MarkupSafe = "3.0.2"
marshmallow = "3.26.1"
mdurl = "0.1.2"
mmh3 = "5.1.0"
monotonic = "1.6"
mpmath = "1.3.0"
multidict = "6.1.0"
multiprocess = "0.70.16"
mypy-extensions = "1.0.0"
nest-asyncio = "1.6.0"
networkx = "3.4.2"
nltk = "3.9.1"
numpy = "1.26.4"
oauthlib = "3.2.2"
onnxruntime = "1.20.1"
openai = "1.63.0"
opentelemetry-api = "1.30.0"
opentelemetry-exporter-otlp-proto-common = "1.30.0"
opentelemetry-exporter-otlp-proto-grpc = "1.30.0"
opentelemetry-instrumentation = "0.51b0"
opentelemetry-instrumentation-asgi = "0.51b0"
opentelemetry-instrumentation-fastapi = "0.51b0"
opentelemetry-proto = "1.30.0"
opentelemetry-sdk = "1.30.0"
opentelemetry-semantic-conventions = "0.51b0"
opentelemetry-util-http = "0.51b0"
orjson = "3.10.15"
overrides = "7.7.0"
packaging = "24.2"
pandas = "2.2.3"
pillow = "11.1.0"
pluggy = "1.5.0"
portalocker = "3.1.1"
posthog = "3.13.0"
propcache = "0.2.1"
protobuf = "5.29.3"
pyarrow = "19.0.0"
pyasn1 = "0.6.1"
pyasn1_modules = "0.4.1"
pydantic = "2.10.6"
pydantic-settings = "2.7.1"
pydantic_core = "2.27.2"
Pygments = "2.19.1"
pypdf = "5.3.0"
PyPika = "0.48.9"
pyproject_hooks = "1.2.0"
pyreadline3 = "3.5.4"
pytest = "8.3.4"
pytest-repeat = "0.9.3"
pytest-xdist = "3.6.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2025.1"
pywin32 = "308"
PyYAML = "6.0.2"
regex = "2024.11.6"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
requests-toolbelt = "1.0.0"
rich = "13.9.4"
rsa = "4.9"
safetensors = "0.5.2"
scikit-learn = "1.6.1"
scipy = "1.15.1"
sentence-transformers = "3.4.1"
sentry-sdk = "2.21.0"
setuptools = "75.8.0"
shellingham = "1.5.4"
six = "1.17.0"
sniffio = "1.3.1"
soupsieve = "2.6"
SQLAlchemy = "2.0.38"
starlette = "0.45.3"
striprtf = "0.0.26"
sympy = "1.13.1"
tabulate = "0.9.0"
tenacity = "9.0.0"
threadpoolctl = "3.5.0"
tiktoken = "0.9.0"
tokenizers = "0.21.0"
torch = "2.6.0"
tqdm = "4.67.1"
transformers = "4.48.3"
typer = "0.15.1"
typing-inspect = "0.9.0"
typing_extensions = "4.12.2"
tzdata = "2025.1"
urllib3 = "2.3.0"
uvicorn = "0.34.0"
watchfiles = "1.0.4"
websocket-client = "1.8.0"
websockets = "14.2"
wrapt = "1.17.2"
xxhash = "3.5.0"
yarl = "1.18.3"
zipp = "3.21.0"
zstandard = "0.23.0"
loguru = "^0.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"